Update `Date.ISO8601FormatStyle` to handle AD and negative years (#403)
* Update `Date.ISO8601FormatStyle` to handle AD and negative years

Also move `BinaryInteger` `numericStringRepresentation` to `FoundationEssentials`

Resolves rdar://122065123
